Added the option to convert output to voltage

+++ hardware-lib/sis-interface/include/SIS3302.hh
@@ -14,7 +14,7 @@ class SIS3302: public SISBase {
 
       int Initialize();
       int ReInitialize();
-      int ReadOutData(std::vector<unsigned short> &outData);
+      int ReadOutData(std::vector<double> &outData);
 
 };
